+++ pymc/sample.py
